refactor: 💡 homepage splitting
